Sutvarkytas admin.py OrderLineAdmin ir OrderLineInLine confliktas su price_product parašyta funkcija get_price_product, kuri išspendė: the price_product field doesn't exist in the OrderLine model, or it's not accessible for some reason
In your OrderLineAdmin class, you've created a method get_price_product to fetch the price_product from the related Product model

after that.

Unable to find the get_price_product field in the OrderLine model.This is because price_product is not a field of OrderLine, but rather a field of the Product model

soved by  get_price_product has been removed from the 'Plokštė' fieldset.